requirements.txt
tqdm
numpy
flask
flask-bootstrap
pillow
nose